Corrected Hourly Analysis and Disabled Synthetic Dataset
## Bayesian Updating with Beta Distribution

### Prior Distribution
A **Beta(1,1)** prior is used for the conversion probability. This is a uniform prior, meaning before observing any data, all conversion rates between 0 and 1 are considered equally likely.

### Posterior Updates
When data is observed (successes and failures), the parameters of the Beta distribution (denoted as \( \alpha \) and \( \beta \)) are updated as follows:

#### For the Ad Group:
\[
\alpha_{\text{ad}} = 1 + \text{adSuccesses}
\]
\[
\beta_{\text{ad}} = 1 + (\text{adTrials} - \text{adSuccesses})
\]

#### For the PSA Group:
\[
\alpha_{\text{psa}} = 1 + \text{psaSuccesses}
\]
\[
\beta_{\text{psa}} = 1 + (\text{psaTrials} - \text{psaSuccesses})
\]

### Posterior Distributions
Using the fact that the Beta distribution is a conjugate prior for the Bernoulli (or binomial) likelihood, the posterior distributions for the conversion probabilities are also Beta distributions:

#### Ad Group Conversion Rate:
\[
p_{\text{ad}} \sim \text{Beta}(\alpha_{\text{ad}}, \beta_{\text{ad}})
\]

#### PSA Group Conversion Rate:
\[
p_{\text{psa}} \sim \text{Beta}(\alpha_{\text{psa}}, \beta_{\text{psa}})
\]